+++ app/auth/register.tsx
@@ -1,220 +1,1 @@
-import React, { useState } from "react";
-import { 
-  View, 
-  Text, 
-  StyleSheet, 
-  TouchableOpacity, 
-  KeyboardAvoidingView,
-  Platform,
-  ScrollView
-} from "react-native";
-import { useRouter } from "expo-router";
-import { useAuthStore } from "@/store/auth-store";
-import { colors } from "@/constants/colors";
-import Input from "@/components/Input";
-import Button from "@/components/Button";
-
-export default function RegisterScreen() {
-  const router = useRouter();
-  const { register, isLoading, error } = useAuthStore();
-  
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  
-  const [usernameError, setUsernameError] = useState("");
-  const [emailError, setEmailError] = useState("");
-  const [passwordError, setPasswordError] = useState("");
-  const [confirmPasswordError, setConfirmPasswordError] = useState("");
-  
-  const validateInputs = () => {
-    let isValid = true;
-    
-    if (!username.trim()) {
-      setUsernameError("El nombre de usuario es obligatorio");
-      isValid = false;
-    } else if (username.length < 3) {
-      setUsernameError("El nombre de usuario debe tener al menos 3 caracteres");
-      isValid = false;
-    } else {
-      setUsernameError("");
-    }
-    
-    if (!email.trim()) {
-      setEmailError("El email es obligatorio");
-      isValid = false;
-    } else if (!/\S+@\S+\.\S+/.test(email)) {
-      setEmailError("Por favor introduce un email válido");
-      isValid = false;
-    } else {
-      setEmailError("");
-    }
-    
-    if (!password.trim()) {
-      setPasswordError("La contraseña es obligatoria");
-      isValid = false;
-    } else if (password.length < 6) {
-      setPasswordError("La contraseña debe tener al menos 6 caracteres");
-      isValid = false;
-    } else {
-      setPasswordError("");
-    }
-    
-    if (!confirmPassword.trim()) {
-      setConfirmPasswordError("Por favor confirma tu contraseña");
-      isValid = false;
-    } else if (password !== confirmPassword) {
-      setConfirmPasswordError("Las contraseñas no coinciden");
-      isValid = false;
-    } else {
-      setConfirmPasswordError("");
-    }
-    
-    return isValid;
-  };
-  
-  const handleRegister = async () => {
-    if (!validateInputs()) return;
-    
-    try {
-      await register(username, email, password);
-      router.replace("/(tabs)");
-    } catch (error) {
-      console.error("Error de registro:", error);
-    }
-  };
-  
-  const handleLogin = () => {
-    router.push("/auth/login");
-  };
-  
-  return (
-    <KeyboardAvoidingView
-      style={styles.container}
-      behavior={Platform.OS === "ios" ? "padding" : "height"}
-      keyboardVerticalOffset={Platform.OS === "ios" ? 64 : 0}
-    >
-      <ScrollView contentContainerStyle={styles.scrollContent}>
-        <View style={styles.header}>
-          <Text style={styles.title}>Crear Cuenta</Text>
-          <Text style={styles.subtitle}>Regístrate para empezar a seguir tus suplementos</Text>
-        </View>
-        
-        <View style={styles.form}>
-          <Input
-            label="Nombre de Usuario"
-            placeholder="Elige un nombre de usuario"
-            value={username}
-            onChangeText={setUsername}
-            autoCapitalize="none"
-            error={usernameError}
-          />
-          
-          <Input
-            label="Email"
-            placeholder="Introduce tu email"
-            value={email}
-            onChangeText={setEmail}
-            keyboardType="email-address"
-            autoCapitalize="none"
-            error={emailError}
-          />
-          
-          <Input
-            label="Contraseña"
-            placeholder="Crea una contraseña"
-            value={password}
-            onChangeText={setPassword}
-            secureTextEntry
-            error={passwordError}
-          />
-          
-          <Input
-            label="Confirmar Contraseña"
-            placeholder="Confirma tu contraseña"
-            value={confirmPassword}
-            onChangeText={setConfirmPassword}
-            secureTextEntry
-            error={confirmPasswordError}
-          />
-          
-          {error && (
-            <Text style={styles.errorText}>{error}</Text>
-          )}
-          
-          <Button
-            title="Crear Cuenta"
-            onPress={handleRegister}
-            loading={isLoading}
-            disabled={isLoading}
-            style={styles.registerButton}
-            fullWidth
-          />
-        </View>
-        
-        <View style={styles.footer}>
-          <Text style={styles.footerText}>¿Ya tienes una cuenta?</Text>
-          <TouchableOpacity onPress={handleLogin}>
-            <Text style={styles.loginText}>Iniciar Sesión</Text>
-          </TouchableOpacity>
-        </View>
-      </ScrollView>
-    </KeyboardAvoidingView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.background,
-  },
-  scrollContent: {
-    flexGrow: 1,
-    padding: 24,
-  },
-  header: {
-    alignItems: "center",
-    marginBottom: 32,
-    marginTop: 40,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: "bold",
-    color: colors.text,
-    marginBottom: 8,
-  },
-  subtitle: {
-    fontSize: 16,
-    color: colors.textSecondary,
-    textAlign: "center",
-  },
-  form: {
-    marginBottom: 24,
-  },
-  registerButton: {
-    marginTop: 16,
-  },
-  footer: {
-    flexDirection: "row",
-    justifyContent: "center",
-    alignItems: "center",
-    marginBottom: 24,
-  },
-  footerText: {
-    fontSize: 14,
-    color: colors.textSecondary,
-    marginRight: 4,
-  },
-  loginText: {
-    fontSize: 14,
-    fontWeight: "600",
-    color: colors.primary,
-  },
-  errorText: {
-    color: colors.danger,
-    fontSize: 14,
-    marginTop: 8,
-    textAlign: "center",
-  },
-});
+export { default } from "@/screens/RegisterScreen";
