+++ package.json
@@ -32,6 +32,7 @@
     "expo-symbols": "~0.4.4",
     "expo-system-ui": "~5.0.6",
     "expo-web-browser": "~14.1.6",
+    "firebase": "^10.14.1",
     "hono": "^4.8.2",
     "lucide-react-native": "^0.475.0",
     "nativewind": "^4.1.23",
